# filter: implements --filter-expr with arbitrary expressions