```
/ # cat /dev/null|cat|tac|head|hexdump|cat

$ sudo ig run trace_exec --paths --filter-expr='exepath != "/bin/cat" && cwd == "/"'
RUNTIME.CONTAINERNAME   COMM         PID        TID ARGS          ERROR
heuristic_bassi         tac       235523     235523 /bin/tac
heuristic_bassi         hexdump   235525     235525 /bin/hexdump
heuristic_bassi         head      235524     235524 /bin/head
```

```
/ # cat /dev/null|head|tail|tac

$ sudo ig run trace_exec --paths --filter-expr='(proc.comm == "cat" || proc.comm == "head") && (runtime.containerImageName == "busybox" || runtime.containerImageName == "ubuntu")'
RUNTIME.CONTAINERNAME   COMM         PID        TID ARGS                     ERROR
happy_cartwright        head      258247     258247 /usr/bin/head
happy_cartwright        cat       258246     258246 /usr/bin/cat /dev/null
```

```
/ # cat /dev/null|head|tail|tac

$ sudo ig run trace_exec --paths --filter-expr='len(proc.comm) == 3 && runtime.containerImageName in ["busybox", "ubuntu"]'
RUNTIME.CONTAINERNAME   COMM         PID        TID ARGS                                                                         ERROR
heuristic_bassi         cat       260243     260243 /bin/cat /dev/null
heuristic_bassi         tac       260246     260246 /bin/tac
happy_cartwright        cat       260253     260253 /usr/bin/cat /dev/null
happy_cartwright        tac       260256     260256 /usr/bin/tac
```


cc @flyth @johananl 

Fixes: #4590

## How to use

TODO

## Testing done

TODO

## TODO

- [x] Test access to sub fields
- [x] Test access to fields of different types
- [x] Don't prefix fields with `data.`
- [ ] ~Find fields at compilation-time instead of evaluation-time~
- [x] Benchmarks