@alban 2025-06-11
I took @flyth 's [PoC](https://github.com/inspektor-gadget/inspektor-gadget/compare/michael/oci/expr) using [expr](https://expr-lang.org/) instead of [Gval](https://github.com/PaesslerAG/gval) and [adapted](6857c043130cb62e038ee7021437d772f811604b) the benchmarks to be able to compare both.

I generated the stats with:
```
METHOD=expr go test -bench=. -run=Benchmark ./pkg/operators/expr/... -count 10 | tee expr.bench
METHOD=gval go test -bench=. -run=Benchmark ./pkg/operators/expr/... -count 10 | tee gval.bench
```

And compared with:
```
$ benchstat gval.bench expr.bench
```

Results:
```
goos: linux
goarch: amd64
pkg: github.com/inspektor-gadget/inspektor-gadget/pkg/operators/expr
cpu: Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz
                                       │  gval.bench  │              expr.bench               │
                                       │    sec/op    │    sec/op     vs base                 │
Expression/string_match_positive-4       1.238µ ±  7%   5.657µ ±  6%  +357.13% (p=0.000 n=10)
Expression/string_match_negative-4       1.258µ ±  7%   5.624µ ±  7%  +347.24% (p=0.000 n=10)
Expression/string_not_match_positive-4   1.252µ ± 17%   5.647µ ± 11%  +351.18% (p=0.000 n=10)
Expression/string_not_match_negative-4   1.233µ ±  9%   5.732µ ± 25%  +364.88% (p=0.000 n=10)
geomean                                  1.245µ         5.665µ        +355.06%
```

This shows that Gval is more than 4 times faster: an operation with expr takes +355.06% more time than an operation with Gval.